Check if hand is full before drawing
GameController.cs: implemented a way of checking if player’s hand is
full before drawing during draw phases. A popup shows up telling the
player to discard a card before drawing a new one.
Issues: it doesn’t actually force the player to discard a card so that
should be changed.

We should also change the way a card is drawn: make the player see new
card to decide if wants to keep it and discard one on hand or discard
the new one and keep hand as it is.

CardPlayer.cs: added function that returns number of cards on hand.