fixing inserting cards on empty spots
Now cards go to first empty spot on hand instead of always going on top
of first card.

CardGame:
Added an array to keep track of available spots on hand and a method to
update when there are free spots.
In DealCards it looks for first empty spot when it’s the player’s hand
and update the handTarget variable.

To fix: If hand is full it is currently doing nothing.

CardObject:
Added variable handPos to keep track of position that the card is in.
Also added methods to update position and to return it.

CardPlayer:
Before calling Turn updates that current spot that card is being taken
from is free.

DiscardPile:
Also updates that current spot is free after card is sold or salvaged.