Game Over Popup
A game over popup shows up when the player runs out of cards.

Game Over prefab:
Created a Game Over popup that has the game over text and “Main Menu”
and “New Game” buttons.

GameController.cs:
Added a private isGameOver method to check if the number of cards in
hand is equal to zero. Call this function on the turn phases that do
not draw cards.

UIManager.cs:
Added the gameOver popup object and a method to set it to visible when
it is called from GameController and condition isGameOver is true.